refactors sigin component
